refactor: fixing requested changes
