+++ packages/demo/src/components/Bandada/index.ts
@@ -1,0 +1,3 @@
+import { Bandada } from "./Bandada";
+
+export default Bandada;
