+++ packages/demo/src/components/Connect/index.ts
@@ -1,0 +1,3 @@
+import { Connect } from "./Connect";
+
+export default Connect;
