+++ packages/demo/src/components/ConnectedIdentity/index.ts
@@ -1,0 +1,3 @@
+import { ConnectedIdentity } from "./ConnectedIdentity";
+
+export default ConnectedIdentity;
