+++ packages/demo/src/components/DisplayProof/index.ts
@@ -1,0 +1,3 @@
+import { DisplayProof } from "./DisplayProof";
+
+export default DisplayProof;
