+++ packages/demo/src/components/GetCommitment/index.ts
@@ -1,0 +1,3 @@
+import { GetCommitment } from "./GetCommitment";
+
+export default GetCommitment;
