+++ packages/demo/src/components/GetMetadata/index.ts
@@ -1,0 +1,3 @@
+import { GetMetadata } from "./GetMetadata";
+
+export default GetMetadata;
