+++ packages/demo/src/components/RateLimitingNullifier/index.ts
@@ -1,0 +1,3 @@
+import { RateLimitingNullifier } from "./RateLimitingNullifier";
+
+export default RateLimitingNullifier;
