+++ packages/demo/src/components/Semaphore/index.ts
@@ -1,0 +1,3 @@
+import { Semaphore } from "./Semaphore";
+
+export default Semaphore;
