+++ packages/demo/src/components/VerifiableCredentials/index.ts
@@ -1,0 +1,3 @@
+import { VerifiableCredentials } from "./VerifiableCredentials";
+
+export default VerifiableCredentials;
