+++ packages/demo/tsconfig.json
@@ -1,9 +1,15 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "baseUrl": ".",
+    "rootDir": ".",
+    "outDir": "dist",
     "experimentalDecorators": true,
     "jsx": "react-jsx",
-    "lib": ["es2020", "dom"]
+    "lib": ["es2020", "dom"],
+    "paths": {
+      "@src/*": ["src/*"]
+    }
   },
   "exclude": ["node_modules", "dist"]
 }
